agent: Swap `/authorize/role/:role` for `/authorize/dekaf`
In addition to avoiding premature generalization, we can now also return relevant metadata in one request, such as ops catalog names and task spec.
